backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
